套餐信息

+++ src/main/java/com/xiliulou/electricity/vo/OwnMemberCardInfoVo.java
@@ -20,7 +20,7 @@ public class OwnMemberCardInfoVo {
     private Long remainingNumber;
 
 
-    private Integer days;
+    private Long days;
     private Integer type;
 
 }
